Enum savefile::SavefileError

#[non_exhaustive]
pub enum SavefileError {
    IncompatibleSchema {
        message: String,
    },
    IOError {
        io_error: Error,
    },
    InvalidUtf8 {
        msg: String,
    },
    MemoryAllocationLayoutError,
    ArrayvecCapacityError {
        msg: String,
    },
    ShortRead,
    CryptographyError,
    SizeOverflow,
    WrongVersion {
        msg: String,
    },
    GeneralError {
        msg: String,
    },
    PoisonedMutex,
}

This object represents an error in deserializing or serializing an item.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

IncompatibleSchema

Error given when the schema stored in a file, does not match the schema given by the data structures in the code, taking into account versions.

Fields of IncompatibleSchema

message: String

A short description of the incompatibility

IOError

Some sort of IO failure. Permissions, broken media etc …

Fields of IOError

io_error: Error

Cause

InvalidUtf8

The binary data which is being deserialized, contained an invalid utf8 sequence where a String was expected. If this occurs, it is either a bug in savefile, a bug in an implementation of Deserialize, Serialize or WithSchema, or a corrupt data file.

Fields of InvalidUtf8

msg: String

descriptive message

MemoryAllocationLayoutError

Unexpected error with regards to memory layout requirements.

ArrayvecCapacityError

An Arrayvec had smaller capacity than the size of the data in the binary file.

Fields of ArrayvecCapacityError

msg: String

Descriptive message

ShortRead

The reader returned fewer bytes than expected

CryptographyError

Cryptographic checksum mismatch. Probably due to a corrupt file.

SizeOverflow

A persisted value of isize or usize was greater than the maximum for the machine. This can happen if a file saved by a 64-bit machine contains an usize or isize which does not fit in a 32 bit word.

WrongVersion

The file does not have a supported version number

Fields of WrongVersion

msg: String

Descriptive message

GeneralError

The file does not have a supported version number

Fields of GeneralError

msg: String

Descriptive message

PoisonedMutex

A poisoned mutex was encountered when traversing the object being saved