Trait savefile::Deserialize

pub trait Deserialize: WithSchema + Sized {
    fn deserialize(
        deserializer: &mut Deserializer<'_>
    ) -> Result<Self, SavefileError>;
}

This trait must be implemented for all data structures you wish to be able to deserialize.

The most convenient way to implement this is to use #[macro_use] extern crate savefile-derive;

and the use #[derive(Deserialize)]

Required methods

fn deserialize(
deserializer: &mut Deserializer<'_>
) -> Result<Self, SavefileError>

Deserialize and return an instance of Self from the given deserializer.